+++ fsl_mrs/tests/test_scripts_fsl_mrsi.py
@@ -41,9 +41,17 @@ def test_fsl_mrsi(tmp_path):
     assert (tmp_path / 'fit_out/qc').exists()
     assert (tmp_path / 'fit_out/uncertainties').exists()
     assert (tmp_path / 'fit_out/concs').exists()
+    assert (tmp_path / 'fit_out/nuisance').exists()
 
     assert (tmp_path / 'fit_out/concs/raw/NAA.nii.gz').exists()
     assert (tmp_path / 'fit_out/concs/molality/NAA.nii.gz').exists()
     assert (tmp_path / 'fit_out/uncertainties/NAA_sd.nii.gz').exists()
     assert (tmp_path / 'fit_out/qc/NAA_snr.nii.gz').exists()
     assert (tmp_path / 'fit_out/fit/fit.nii.gz').exists()
+
+    assert (tmp_path / 'fit_out/nuisance/p0.nii.gz').exists()
+    assert (tmp_path / 'fit_out/nuisance/p1.nii.gz').exists()
+    assert (tmp_path / 'fit_out/nuisance/shift_group0.nii.gz').exists()
+    assert (tmp_path / 'fit_out/nuisance/combined_lw_group0.nii.gz').exists()
+    assert (tmp_path / 'fit_out/nuisance/gamma_group0.nii.gz').exists()
+    assert (tmp_path / 'fit_out/nuisance/sigma_group0.nii.gz').exists()