+++ fsl_mrs/core/mrsi.py
@@ -27,7 +27,7 @@ class MRSI(object):
 
         # process H2O
         if H2O is None:
-            H2O = np.full(FID.shape, None)
+            H2O = np.full(FID.shape[:3], None)
         elif H2O.shape != FID.shape:
             raise ValueError('H2O must be None or numpy array '
                              'of the same shape as FID.')
@@ -208,9 +208,12 @@ class MRSI(object):
         as a single MRS object.
         '''
         FID = misc.volume_to_list(self.data, self.mask)
-        H2O = misc.volume_to_list(self.H2O, self.mask)
         FID = sum(FID) / len(FID)
-        H2O = sum(H2O) / len(H2O)
+        if not np.array_equal(self.H2O, np.full(self.data.shape[:3], None)):
+            H2O = misc.volume_to_list(self.H2O, self.mask)
+            H2O = sum(H2O) / len(H2O)
+        else:
+            H2O = None
 
         mrs_out = MRS(FID=FID,
                       header=self.header,